+++ src/com/intellij/idea/plugin/hybris/codeInspection/rule/impex/AbstractImpExInvalidClassReferenceModifierInspection.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -47,6 +47,7 @@ abstract class AbstractImpExInvalidClassReferenceModifierInspection(
 
         override fun visitAnyAttributeValue(element: ImpexAnyAttributeValue) {
             if (modifier.modifierName != element.anyAttributeName?.text) return
+            if (element.text.startsWith("$")) return
             val reference = element.reference
 
             if (reference == null) unresolvedReference(element)