+++ src/com/intellij/idea/plugin/hybris/impex/lang/folding/util/ImpExSmartFoldingBlocksFilter.kt
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,11 +19,13 @@
 package com.intellij.idea.plugin.hybris.impex.lang.folding.util
 
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexAttribute
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexParameters
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexUserRightsPermissionValue
 import com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils
 import com.intellij.openapi.components.Service
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
 
 @Service
 class ImpExSmartFoldingBlocksFilter : AbstractImpExFoldingFilter() {
@@ -31,10 +33,10 @@ class ImpExSmartFoldingBlocksFilter : AbstractImpExFoldingFilter() {
     override fun isFoldable(element: PsiElement) = isSupportedType(element)
         && (ImpexPsiUtils.isLineBreak(element) || isNotBlankPlaceholder(element))
 
-    private fun isSupportedType(element: PsiElement) = element is ImpexAttribute
-        || element is ImpexParameters
+    private fun isSupportedType(element: PsiElement) = element is ImpexParameters
         || element is ImpexUserRightsPermissionValue
         || ImpexPsiUtils.isLineBreak(element)
+        || (element is ImpexAttribute && PsiTreeUtil.findChildOfType(element, ImpexMacroUsageDec::class.java) == null)
 
     private fun isNotBlankPlaceholder(element: PsiElement) = ImpExSmartFoldingPlaceholderBuilder.getInstance().getPlaceholder(element)
         .isNotBlank()