+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Introduced ImpEx-like custom CSV Format for `Data Edit` mode [#1393](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1393)
 - Added new configurable setting _First row is header_ for `Data Edit` mode [#1394](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1394)
 - Added new configurable setting _Trim whitespace_ for `Data Edit` mode [#1395](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1395)
+- Added folding for macros usages in the header line [#1401](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1401)
 
 ### `Access Control Lists` enhancements
 - Create Type/attributes references only within `Target` column [#1400](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1400)